feature: reduce redunant templates code
I did this by making a generic builder that that work for the rubric and
the templates. We can wire it up to RubricsMain later
