+++ frontend/src/components/index.ts
@@ -21,3 +21,4 @@ export * from "./contextSelectors/ActiveCourseSelection.tsx";
 export * from "./contextSelectors/CourseSelectionMenu.tsx";
 export * from "./contextSelectors/AssignmentSelectionMenu.tsx";
 export * from "./PopUp.tsx";
+export * from "./layout/GenericBuilder.tsx";
