+++ frontend/src/components/modals/Dialog.tsx
@@ -17,9 +17,11 @@ export function Dialog({ isOpen, onClose, title, children }: DialogProps) {
 
   return createPortal(
     <div className="scroll-auto fixed z-80 inset-0 bg-black bg-opacity-75 flex justify-center items-center">
-      <div className="bg-gray-700 p-6 rounded shadow-lg relative w-1/3 max-w-4xl">
-        <h2 className="text-xl text-white font-semibold j">{title}</h2>
-        <div className={"text-gray-100"}>{children}</div>
+      <div className="bg-gray-700 px-4 py-6 rounded shadow-lg relative w-2/5 max-w-4xl">
+        <h2 className="text-xl text-white font-semibold">{title}</h2>
+        <div className={"text-gray-100 flex items-center justify-center "}>
+          {children}
+        </div>
         <button
           className="absolute top-2 right-2 text-2xl text-black font-bold hover:text-red-600 hover:scale-110 transition-colors ease-in-out duration-300"
           onClick={onClose}
@@ -28,7 +30,7 @@ export function Dialog({ isOpen, onClose, title, children }: DialogProps) {
         </button>
       </div>
     </div>,
-    document.getElementById("portal-root") as HTMLElement, // render outside the primary DOM to ensure modals are in
+    document.getElementById("portal-root") as HTMLElement // render outside the primary DOM to ensure modals are in
     // front of everything else
   );
 }
